rust-server-cleanup(template-split,non_mechanical_provable): split chat template helpers
